// utils/types.ts
export interface ProductI extends Omit<Product, "images" | "slideColors"> {
  images: UploadResponse[];
  category: Record<"name" | "id", string>[];
  slideColors: Record<"color" | "backgroundColor", string>[];
  sizes: string[]; // NEW: Sizes
  keyFeatures: string[]; // NEW: Key Features
  careInstructions: string; // NEW: Care Instructions
  specifications: { label: string; value: string }[]; // NEW: Specifications

}
// ...
export interface LoadingI<T> {
  type: T;
  isLoading: boolean;
}

export type StateUpdateType = "add" | "update" | "delete";